complete filter and search functionality

+++ lib/screens/CartPage/model/cart_model.dart
@@ -1,9 +1,14 @@
-class CartModel{
+class CartModel {
   String title;
   String image;
   int quantity;
-  int price;
-  int totalPrice;
+  double price;
+  double totalPrice;
 
-  CartModel({required this.title , required this.image , required this.quantity, required this.price,required this.totalPrice});
+  CartModel(
+      {required this.title,
+      required this.image,
+      required this.quantity,
+      required this.price,
+      required this.totalPrice});
 }
